Prepare analytics containment API
