+++ Sources/Analytics/UserInterface/ContainerPageViewTracking.swift
@@ -1,0 +1,11 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import UIKit
+
+public protocol ContainerPageViewTracking {
+    var activeChildViewControllers: [UIViewController] { get }
+}
