+++ Sources/Analytics/UserInterface/PageViewTracking.swift
@@ -21,18 +21,3 @@ public extension PageViewTracking {
         true
     }
 }
-
-extension UIViewController {
-    static func setupTracking() {
-        method_exchangeImplementations(
-            class_getInstanceMethod(Self.self, #selector(viewDidAppear(_:)))!,
-            class_getInstanceMethod(Self.self, #selector(swizzledViewDidAppear(_:)))!
-        )
-    }
-
-    @objc func swizzledViewDidAppear(_ animated: Bool) {
-        swizzledViewDidAppear(animated)
-        guard let trackedViewController = self as? PageViewTracking, trackedViewController.isTrackedAutomatically else { return }
-        Analytics.shared.sendPageView(title: trackedViewController.pageTitle, levels: trackedViewController.pageLevels)
-    }
-}
